app/requirements.txt
Flask
Flask-Cors
gunicorn
google-generativeai
Pillow
numpy
faiss-cpu
rich
psutil
sentence-transformers
torch
pynput
